+++ backend/src/main/java/com/back/fairytale/domain/user/service/AuthService.java
@@ -4,9 +4,9 @@ package com.back.fairytale.domain.user.service;
 import com.back.fairytale.domain.user.dto.TokenPairDto;
 import com.back.fairytale.domain.user.entity.User;
 import com.back.fairytale.domain.user.repository.UserRepository;
-import com.back.fairytale.global.security.JWTProvider;
-import com.back.fairytale.global.security.LogoutService;
-import com.back.fairytale.global.security.UserTokenService;
+import com.back.fairytale.global.security.jwt.JWTProvider;
+import com.back.fairytale.global.security.port.LogoutService;
+import com.back.fairytale.global.security.port.UserTokenService;
 import jakarta.servlet.http.Cookie;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
