+++ backend/src/main/java/com/back/fairytale/domain/user/service/OAuth2UserManagerImpl.java
@@ -4,9 +4,9 @@ import com.back.fairytale.domain.user.entity.User;
 import com.back.fairytale.domain.user.enums.IsDeleted;
 import com.back.fairytale.domain.user.enums.Role;
 import com.back.fairytale.domain.user.repository.UserRepository;
-import com.back.fairytale.global.security.JWTProvider;
-import com.back.fairytale.global.security.OAuth2UserInfo;
-import com.back.fairytale.global.security.OAuth2UserManager;
+import com.back.fairytale.global.security.jwt.JWTProvider;
+import com.back.fairytale.global.security.oauth2.OAuth2UserInfo;
+import com.back.fairytale.global.security.port.OAuth2UserManager;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
