+++ backend/src/main/java/com/back/fairytale/global/security/oauth2/CustomOAuth2UserService.java
@@ -1,5 +1,7 @@
-package com.back.fairytale.global.security;
+package com.back.fairytale.global.security.oauth2;
 
+import com.back.fairytale.global.security.CustomOAuth2User;
+import com.back.fairytale.global.security.port.OAuth2UserManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
