+++ backend/src/main/java/com/back/fairytale/global/security/port/UserTokenService.java
@@ -1,4 +1,4 @@
-package com.back.fairytale.global.security;
+package com.back.fairytale.global.security.port;
 
 public interface UserTokenService {
     String getUserToken(Long userId);
